[IsolationSession] Refactor runner into per-responsibility module files
