+++ src/wxc_common/src/isolation_session/mod.rs
@@ -1,0 +1,39 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+//! IsolationSession backend — executes scripts in an isolated Windows
+//! session via the in-proc `Windows.AI.IsolationSession` `IsoSessionOps`
+//! API. `IsolationSessionRunner` is the only externally-reachable type;
+//! the granular lifecycle wrapper (`IsolationSessionManager`) and helpers
+//! are module-private.
+//!
+//! Trait impls split by lifecycle shape:
+//! - `one_shot`: `ScriptRunner` — register → provision → start → exec →
+//!   stop → deprovision in a single process.
+//! - `state_aware`: `StatefulSandboxBackend` — per-phase methods called
+//!   across multiple `wxc-exec` invocations by an external orchestrator.
+
+mod error;
+mod folder_sharing;
+mod manager;
+mod one_shot;
+mod policy;
+mod process_options;
+mod protected_paths_filter;
+mod state_aware;
+
+/// Stateless marker type. Trait impls live in `one_shot` (`ScriptRunner`)
+/// and `state_aware` (`StatefulSandboxBackend`).
+pub struct IsolationSessionRunner;
+
+impl IsolationSessionRunner {
+    pub fn new() -> Self {
+        Self
+    }
+}
+
+impl Default for IsolationSessionRunner {
+    fn default() -> Self {
+        Self::new()
+    }
+}
